+++ src/plugins/atlanta/index.ts
@@ -1,0 +1,2 @@
+export { AtlantaCrimePlugin, type AtlantaCrimePluginConfig } from './atlanta-crime.plugin';
+export { AtlantaTrafficPlugin, type AtlantaTrafficPluginConfig } from './atlanta-traffic.plugin';